requirements.txt
Brotli==1.0.9
click==8.0.1
colorama==0.4.4
dash==1.21.0
dash-bootstrap-components==0.12.2
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
numpy==1.21.1
pandas==1.1.5
plotly==5.1.0
protobuf==3.17.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
tenacity==8.0.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
